Docker/10RobotFramework/server/application/requirements.txt
kafka-python==2.0.2
boto3==1.34.0
botocore>=1.29.0
requests==2.31.0
